Fix add-in repo creation issue
Due to a bug, deleting an add-in source, breaks the releases that have been
created with that source. The problem is that the add-in that has been built
is deleted, so it can't be copied to the packages directory. As a workaround,
we now detect if a release doesn't have a source, and in that case we
take the add-in binary from the original packages directory.